Move update sequencer out into own file.
